Handled public groups and joined groups in Groups Component instead of using separate components
